Move classes for metric calculation to separate metric package
