+++ src/main/java/de/uhd/ifi/se/decision/management/jira/filtering/FilteringManager.java
@@ -22,8 +22,8 @@ import de.uhd.ifi.se.decision.management.jira.model.KnowledgeStatus;
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
 import de.uhd.ifi.se.decision.management.jira.model.Link;
 import de.uhd.ifi.se.decision.management.jira.model.LinkType;
-import de.uhd.ifi.se.decision.management.jira.quality.completeness.DefinitionOfDone;
-import de.uhd.ifi.se.decision.management.jira.quality.completeness.KnowledgeElementCheck;
+import de.uhd.ifi.se.decision.management.jira.quality.DefinitionOfDone;
+import de.uhd.ifi.se.decision.management.jira.quality.KnowledgeElementCheck;
 
 /**
  * Filters the {@link KnowledgeGraph}. The filter criteria are specified in the
