+++ src/main/java/de/uhd/ifi/se/decision/management/jira/metric/CommentMetricCalculator.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.quality.generalmetrics;
+package de.uhd.ifi.se.decision.management.jira.metric;
 
 import java.util.ArrayList;
 import java.util.HashMap;
