+++ src/main/java/de/uhd/ifi/se/decision/management/jira/metric/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Provides classes for the <b>calculation of metrics on the knowledge graph
+ * data structure</b>. The metrics are presented in the knowledge dashboard.
+ * 
+ * @see <a href=
+ *      "https://github.com/cures-hub/cures-condec-jira/blob/master/doc/features/dashboard.md">
+ *      README for the Knowledge Dashboard</a>
+ */
+package de.uhd.ifi.se.decision.management.jira.metric;
