+++ src/main/java/de/uhd/ifi/se/decision/management/jira/quality/DefinitionOfDoneChecker.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.quality.completeness;
+package de.uhd.ifi.se.decision.management.jira.quality;
 
 import static java.util.Map.entry;
