+++ src/main/java/de/uhd/ifi/se/decision/management/jira/quality/KnowledgeElementCheck.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.quality.completeness;
+package de.uhd.ifi.se.decision.management.jira.quality;
 
 import java.util.List;
