+++ src/main/java/de/uhd/ifi/se/decision/management/jira/rest/DashboardRest.java
@@ -11,9 +11,9 @@ import com.atlassian.jira.user.ApplicationUser;
 import com.google.common.collect.ImmutableMap;
 
 import de.uhd.ifi.se.decision.management.jira.filtering.FilterSettings;
-import de.uhd.ifi.se.decision.management.jira.quality.completeness.RationaleCompletenessCalculator;
-import de.uhd.ifi.se.decision.management.jira.quality.completeness.RationaleCoverageCalculator;
-import de.uhd.ifi.se.decision.management.jira.quality.generalmetrics.GeneralMetricCalculator;
+import de.uhd.ifi.se.decision.management.jira.metric.GeneralMetricCalculator;
+import de.uhd.ifi.se.decision.management.jira.metric.RationaleCompletenessCalculator;
+import de.uhd.ifi.se.decision.management.jira.metric.RationaleCoverageCalculator;
 
 /**
  * REST resource for dashboards
