+++ src/main/java/de/uhd/ifi/se/decision/management/jira/view/treeviewer/TreeViewerNode.java
@@ -15,7 +15,7 @@ import de.uhd.ifi.se.decision.management.jira.model.KnowledgeElement;
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
 import de.uhd.ifi.se.decision.management.jira.model.Link;
 import de.uhd.ifi.se.decision.management.jira.model.PartOfJiraIssueText;
-import de.uhd.ifi.se.decision.management.jira.quality.completeness.DefinitionOfDoneChecker;
+import de.uhd.ifi.se.decision.management.jira.quality.DefinitionOfDoneChecker;
 
 /**
  * Model class for Tree Viewer nodes
