+++ src/main/java/de/uhd/ifi/se/decision/management/jira/view/vis/VisTimeLineNode.java
@@ -8,7 +8,7 @@ import javax.xml.bind.annotation.XmlElement;
 import de.uhd.ifi.se.decision.management.jira.filtering.FilterSettings;
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeElement;
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
-import de.uhd.ifi.se.decision.management.jira.quality.completeness.DefinitionOfDoneChecker;
+import de.uhd.ifi.se.decision.management.jira.quality.DefinitionOfDoneChecker;
 
 public class VisTimeLineNode {
