+++ src/test/java/de/uhd/ifi/se/decision/management/jira/metric/TestRationaleCoverageCalculator.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.quality.completeness;
+package de.uhd.ifi.se.decision.management.jira.metric;
 
 import static org.junit.Assert.assertEquals;
